Refactor UR functions: Move UR_ prefixed functions from UR10_Server_functions.py to new UR_Common_functions.py for better code organization and maintainability. Update scanner image handling in UR20_Server_functions.py to support multiple scanner statuses.
